Extract Notification Settings into Reusable Component
